Added working drag and drop feature.
Created basic 2d squares as placeholders for the commands and made basic event squares for where the commands go. Finally, created a system to save the players position when they press S (Creating a clone) so that the UI toggles on when they're within close proximity to the saved clone position - Rudimentary for now, my next steps are going to be refining this system.